blockchain tests
Conflicts:
	test/block.cpp
